+++ scripts/build-ultra.mjs
@@ -1,0 +1,9 @@
+import { build } from 'vite';
+
+import viteConfig from '../vite.config.mjs';
+import { compressDist } from './compress-dist.mjs';
+import { postbuildCopyStatic } from './postbuild-copy.mjs';
+
+await build(viteConfig);
+postbuildCopyStatic();
+await compressDist();